pyproject.toml
[tool.poetry]
name = "heroic-rl"
version = "1.0.0"
description = "Heroic RL framework"
authors = ["Dimitrije Radojević <dimitrijer@nordeus.com>",
           "Michał Warchalski <michalw@nordeus.com>"]
repository = "https://github.com/Nordeus/heroic-rl"
readme = "README.md"
license = "GPL-3.0-only"
include = ["decks.csv", "images"]

[tool.poetry.scripts]
heroic-rl = "heroic_rl.console:run"

[tool.poetry.dependencies]
python = "^3.6.1"
Flask = "^1.1"
requests = "^2.22"
gym = "^0.12"
joblib = "^0.13"
mpi4py = "^3.0.2"
numpy = "^1.16"
cloudpickle = "^1.2"
click = "^7.0"
pyyaml = "^5.3.1"
psutil = "^5.6"
# Pin down TF to v2.0 because of CUDA 10.0
tensorflow-gpu = { version = "2.0.1", optional = true }

[tool.poetry.dev-dependencies]
pre-commit = "^2"
flake8 = "^3"
tensorflow = "2.0.1"
flake8-bugbear = "^20"
black = "=19.10b0"
isort = "^4"
seed-isort-config = "^2"

[tool.poetry.extras]
gpu = ["tensorflow-gpu"]

[tool.black]
target-version = ['py36', 'py37']
include = "heroic_rl"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
known_third_party = ["click", "flask", "gym", "joblib", "mpi4py", "numpy", "requests", "scipy", "tensorflow", "termcolor", "werkzeug", "yaml"]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
